fix: 修复合并 v1.14.0-alpha.28-reF1nd 后的编译错误和特性兼容
- 修复 adapter/inbound.go 重复类型/方法声明
- 修复 clashapi/server.go 重复路由挂载和结构体字段
- 修复 interrupt/conn.go 重复 ReadPacket/WritePacket 方法
- 修复 urltest.go 重复 unified_delay 代码块
- 恢复 rule_set_remote.go 响应超限检查和容错启动
- 恢复 provider/remote.go 容错启动（fetch 失败不阻塞进程）
- 添加 Smart/GeoX option 类型定义和 ExperimentalOptions 字段
- 添加 TypeSmart 常量和 OutboundGroup Hidden/Icon 接口实现
- 添加 httpclient.Manager.LookupDetour 方法
- 移除已废弃的 HintUnreachable/RegisterNetworkResetCallback 接口
- 清理 route/network.go 遗留死代码字段

+++ experimental/smart/service.go
@@ -230,7 +230,7 @@ func (s *Service) initCollector() error {
 	}
 	path = filemanager.BasePath(s.ctx, path)
 
-	dc, err := lightgbm.NewDataCollector(path, opts.SizeLimitMB, s.logger)
+	dc, err := lightgbm.NewDataCollector(path, int64(opts.SizeLimitMB), s.logger)
 	if err != nil {
 		s.logger.Warn("smart collector: init failed: ", err)
 		return err
